+++ src/main/java/org/csanchez/jenkins/plugins/kubernetes/KubernetesCloud.java
@@ -188,9 +188,7 @@ public class KubernetesCloud extends Cloud {
      */
     @Nonnull
     public List<PodTemplate> getAllTemplates() {
-        List<PodTemplate> podTemplates = new ArrayList<>(PodTemplateMap.get().getTemplates(this));
-        podTemplates.addAll(templates);
-        return Collections.unmodifiableList(podTemplates);
+        return PodTemplateSource.getAll(this);
     }
 
     @DataBoundSetter
@@ -392,16 +390,16 @@ public class KubernetesCloud extends Cloud {
     @Override
     public synchronized Collection<NodeProvisioner.PlannedNode> provision(@CheckForNull final Label label, final int excessWorkload) {
         try {
-
-            LOGGER.log(Level.INFO, "Excess workload after pending Spot instances: " + excessWorkload);
+            Set<String> allInProvisioning = InProvisioning.getAllInProvisioning(label);
+            LOGGER.log(Level.FINE, () -> "In provisioning : " + allInProvisioning);
+            int toBeProvisioned = Math.max(0, excessWorkload - allInProvisioning.size());
+            LOGGER.log(Level.INFO, "Excess workload after pending Kubernetes agents: " + toBeProvisioned);
 
             List<NodeProvisioner.PlannedNode> r = new ArrayList<NodeProvisioner.PlannedNode>();
 
-            ArrayList<PodTemplate> templates = getMatchingTemplates(label);
-
-            for (PodTemplate t: templates) {
+            for (PodTemplate t: getTemplatesFor(label)) {
                 LOGGER.log(Level.INFO, "Template: " + t.getDisplayName());
-                for (int i = 1; i <= excessWorkload; i++) {
+                for (int i = 1; i <= toBeProvisioned; i++) {
                     if (!addProvisionedSlave(t, label)) {
                         break;
                     }
@@ -491,16 +489,20 @@ public class KubernetesCloud extends Cloud {
      * Gets all PodTemplates that have the matching {@link Label}.
      * @param label label to look for in templates
      * @return list of matching templates
+     * @deprecated Use {@link #getTemplatesFor(Label)} instead.
      */
+    @Deprecated
     public ArrayList<PodTemplate> getMatchingTemplates(@CheckForNull Label label) {
-        ArrayList<PodTemplate> podList = new ArrayList<PodTemplate>();
-        List<PodTemplate> podTemplates = getAllTemplates();
-        for (PodTemplate t : podTemplates) {
-            if ((label == null && t.getNodeUsageMode() == Node.Mode.NORMAL) || (label != null && label.matches(t.getLabelSet()))) {
-                podList.add(t);
-            }
-        }
-        return podList;
+        return new ArrayList<>(getTemplatesFor(label));
+    }
+
+    /**
+     * Gets all PodTemplates that have the matching {@link Label}.
+     * @param label label to look for in templates
+     * @return list of matching templates
+     */
+    public List<PodTemplate> getTemplatesFor(@CheckForNull Label label) {
+        return PodTemplateFilter.applyAll(this, getAllTemplates(), label);
     }
 
     /**
@@ -635,4 +637,12 @@ public class KubernetesCloud extends Cloud {
         return this;
     }
 
+    @Extension
+    public static class PodTemplateSourceImpl extends PodTemplateSource {
+        @Nonnull
+        @Override
+        public List<PodTemplate> getList(@Nonnull KubernetesCloud cloud) {
+            return cloud.getTemplates();
+        }
+    }
 }